feat: apply alert history modal, resolved qa issue (#5471)
* fix: fixed event rule mount bug


* refactor: changed file, type name


* feat: apply alert history modal


* chore: update translations


---------
