+++ apps/web/src/schema/alert-manager/alert/api-verbs/history.ts
@@ -1,3 +1,4 @@
 export interface AlertHistoryParameters {
     alert_id: string;
+    include_details?: boolean;
 }
