+++ apps/web/src/schema/alert-manager/alert/constants.ts
@@ -22,10 +22,16 @@ export const ALERT_TRIGGERED_TYPE = {
     WEBHOOK: 'WEBHOOK',
 } as const;
 
-export const ALERT_EVENT_ACTION = {
+export const ALERT_HISTORY_ACTION = {
     TRIGGERED: 'TRIGGERED',
     ACKNOWLEDGED: 'ACKNOWLEDGED',
     RESOLVED: 'RESOLVED',
-    NOTIFIED: 'NOTIFIED',
+    NOTIFIED_SUCCESS: 'NOTIFIED_SUCCESS',
+    NOTIFIED_FAILURE: 'NOTIFIED_FAILURE',
     EVENT_PUSHED: 'EVENT_PUSHED',
-};
+} as const;
+
+export const ALERT_HISTORY_NOTIFICATION_STATE = {
+    SUCCESS: 'SUCCESS',
+    FAILURE: 'FAILURE',
+} as const;
