+++ apps/web/src/services/alert-manager/components/AlertDetailTabsTimeline.vue
@@ -6,25 +6,23 @@ import {
 
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import {
-    PButton, PButtonModal, PCodeEditor, PToolbox, PHeading, PHeadingLayout, PDataLoader, PDivider, PSelectStatus, PTextButton,
+    PButton, PToolbox, PHeading, PHeadingLayout, PDataLoader, PDivider, PSelectStatus, PTextButton,
 } from '@cloudforet/mirinae';
 
 import type { ListResponse } from '@/schema/_common/api-verbs/list';
 import type { AlertHistoryParameters } from '@/schema/alert-manager/alert/api-verbs/history';
-import { ALERT_EVENT_ACTION } from '@/schema/alert-manager/alert/constants';
-import type { AlertModel, AlertEventModel } from '@/schema/alert-manager/alert/model';
-import type { AlertEventActionType } from '@/schema/alert-manager/alert/type';
+import { ALERT_HISTORY_ACTION } from '@/schema/alert-manager/alert/constants';
+import type { AlertModel, AlertHistoryModel } from '@/schema/alert-manager/alert/model';
+import type { AlertHistoryActionType } from '@/schema/alert-manager/alert/type';
 import { i18n } from '@/translations';
 
 import { useAllReferenceStore } from '@/store/reference/all-reference-store';
 import type { WebhookReferenceMap } from '@/store/reference/webhook-reference-store';
 
-import { copyAnyData } from '@/lib/helper/copy-helper';
-import { showSuccessMessage } from '@/lib/helper/notice-alert-helper';
-
 import VerticalTimelineItem from '@/common/components/vertical-timeline/VerticalTimelineItem.vue';
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
+import AlertDetailTabsTimelineModal from '@/services/alert-manager/components/AlertDetailTabsTimelineModal.vue';
 import { useAlertDetailPageStore } from '@/services/alert-manager/stores/alert-detail-page-store';
 import type { AlertFilterType } from '@/services/alert-manager/types/alert-manager-type';
 
@@ -46,8 +44,8 @@ const storeState = reactive({
 });
 const state = reactive({
     loading: true,
-    historyList: [] as AlertEventModel[],
-    slicedHistoryList: computed<AlertEventModel[]>(() => {
+    historyList: [] as AlertHistoryModel[],
+    slicedHistoryList: computed<AlertHistoryModel[]>(() => {
         let _list = state.historyList;
         if (filterState.searchText) {
             _list = state.historyList.filter((item) => item.description.toLowerCase().includes(filterState.searchText.toLowerCase()));
@@ -55,24 +53,28 @@ const state = reactive({
             _list = state.historyList;
         }
         if (filterState.selectedAction !== 'ALL') {
-            _list = state.historyList.filter((item) => item.action === filterState.selectedAction);
+            if (filterState.selectedAction === 'NOTIFIED') {
+                _list = state.historyList.filter((item) => item.action === ALERT_HISTORY_ACTION.NOTIFIED_FAILURE || item.action === ALERT_HISTORY_ACTION.NOTIFIED_SUCCESS);
+            } else {
+                _list = state.historyList.filter((item) => item.action === filterState.selectedAction);
+            }
         }
         return _list.slice(0, state.pageStart * state.pageLimit);
     }),
     pageStart: 1,
     pageLimit: 10,
-    selectedItem: {} as any,
+    selectedItem: {} as AlertHistoryModel,
     modalVisible: false,
-    isAlertVisible: false,
+    modalType: ALERT_HISTORY_ACTION.EVENT_PUSHED as AlertHistoryActionType,
 });
 const filterState = reactive({
     statusFields: computed<AlertFilterType[]>(() => ([
         { label: i18n.t('ALERT_MANAGER.ALERTS.ALL'), name: 'ALL' },
-        { label: i18n.t('ALERT_MANAGER.ALERTS.TRIGGERED'), name: ALERT_EVENT_ACTION.TRIGGERED },
-        { label: i18n.t('ALERT_MANAGER.ALERTS.ACKNOWLEDGED'), name: ALERT_EVENT_ACTION.ACKNOWLEDGED },
-        { label: i18n.t('ALERT_MANAGER.ALERTS.RESOLVED'), name: ALERT_EVENT_ACTION.RESOLVED },
-        { label: i18n.t('ALERT_MANAGER.ALERTS.NOTIFIED'), name: ALERT_EVENT_ACTION.NOTIFIED },
-        { label: i18n.t('ALERT_MANAGER.ALERTS.EVENT_PUSHED'), name: ALERT_EVENT_ACTION.EVENT_PUSHED },
+        { label: i18n.t('ALERT_MANAGER.ALERTS.TRIGGERED'), name: ALERT_HISTORY_ACTION.TRIGGERED },
+        { label: i18n.t('ALERT_MANAGER.ALERTS.ACKNOWLEDGED'), name: ALERT_HISTORY_ACTION.ACKNOWLEDGED },
+        { label: i18n.t('ALERT_MANAGER.ALERTS.RESOLVED'), name: ALERT_HISTORY_ACTION.RESOLVED },
+        { label: i18n.t('ALERT_MANAGER.ALERTS.NOTIFIED'), name: 'NOTIFIED' },
+        { label: i18n.t('ALERT_MANAGER.ALERTS.EVENT_PUSHED'), name: ALERT_HISTORY_ACTION.EVENT_PUSHED },
     ])),
     selectedAction: 'ALL',
     searchText: '',
@@ -84,24 +86,22 @@ const getCreatedByNames = (createdBy: string): string => {
     }
     return createdBy;
 };
-const getItemInfo = (item: AlertEventActionType): HistoryItemInfo => {
+const getItemInfo = (item: AlertHistoryActionType): HistoryItemInfo => {
     let styleType: string|undefined;
-    if (item === ALERT_EVENT_ACTION.TRIGGERED) styleType = 'red';
-    if (item === ALERT_EVENT_ACTION.ACKNOWLEDGED) styleType = 'violet';
-    if (item === ALERT_EVENT_ACTION.RESOLVED) styleType = 'green';
-    if (item === ALERT_EVENT_ACTION.NOTIFIED) styleType = 'yellow';
-    if (item === ALERT_EVENT_ACTION.EVENT_PUSHED) styleType = 'gray';
+    if (item === ALERT_HISTORY_ACTION.TRIGGERED) styleType = 'red';
+    if (item === ALERT_HISTORY_ACTION.ACKNOWLEDGED) styleType = 'violet';
+    if (item === ALERT_HISTORY_ACTION.RESOLVED) styleType = 'green';
+    if (item === ALERT_HISTORY_ACTION.NOTIFIED_FAILURE || item === ALERT_HISTORY_ACTION.NOTIFIED_SUCCESS) styleType = 'yellow';
+    if (item === ALERT_HISTORY_ACTION.EVENT_PUSHED) styleType = 'gray';
     return {
         title: item.toLowerCase().replace(/_/g, ' ').replace(/\b\w/g, (char) => char.toUpperCase()),
         styleType,
     };
 };
 
-const handleCloseModal = () => {
-    state.modalVisible = false;
-};
-const handleClickHistoryItem = (item: AlertEventModel) => {
+const handleClickHistoryItem = (item: AlertHistoryModel, action: AlertHistoryActionType) => {
     state.modalVisible = true;
+    state.modalType = action;
     state.selectedItem = item;
 };
 const handleSelectAction = (value: string) => {
@@ -116,16 +116,12 @@ const handleClickShowMore = async () => {
     state.pageStart += 1;
 };
 
-const handleClickCopy = () => {
-    copyAnyData(state.selectedItem);
-    showSuccessMessage(i18n.t('ALERT_MANAGER.ALERTS.COPIED'), '');
-};
-
 const fetchHistoryList = async () => {
     state.loading = true;
     try {
-        const { results } = await SpaceConnector.clientV2.alertManager.alert.history<AlertHistoryParameters, ListResponse<AlertEventModel>>({
+        const { results } = await SpaceConnector.clientV2.alertManager.alert.history<AlertHistoryParameters, ListResponse<AlertHistoryModel>>({
             alert_id: storeState.alertInfo.alert_id,
+            include_details: true,
         });
         state.historyList = results || [];
     } catch (e: any) {
@@ -195,10 +191,12 @@ watch(() => storeState.alertInfo, async (alertInfo) => {
                                 {{ $t('ALERT_MANAGER.ALERTS.CREATED_BY', { user: getCreatedByNames(item.created_by) }) }}
                             </span>
                         </template>
-                        <p-text-button v-if="item.action === ALERT_EVENT_ACTION.EVENT_PUSHED"
+                        <p-text-button v-if="item.action === ALERT_HISTORY_ACTION.EVENT_PUSHED
+                                           || item.action === ALERT_HISTORY_ACTION.NOTIFIED_FAILURE
+                                           || item.action === ALERT_HISTORY_ACTION.NOTIFIED_SUCCESS"
                                        style-type="highlight"
                                        size="md"
-                                       @click="handleClickHistoryItem(item)"
+                                       @click="handleClickHistoryItem(item, item.action)"
                         >
                             {{ item.description }}
                         </p-text-button>
@@ -227,36 +225,11 @@ watch(() => storeState.alertInfo, async (alertInfo) => {
                 {{ $t('ALERT_MANAGER.SHOW_MORE') }}
             </p-button>
         </div>
-        <p-button-modal v-if="state.modalVisible"
-                        :header-title="$t('ALERT_MANAGER.ALERTS.EVENT_DETAILS')"
-                        size="lg"
-                        hide-footer-close-button
-                        :visible.sync="state.modalVisible"
-                        @confirm="handleCloseModal"
-        >
-            <template #body>
-                <div class="event-detail-modal-content">
-                    <p-code-editor :code="state.selectedItem"
-                                   class="code-block"
-                                   read-only
-                                   folded
-                    />
-                </div>
-            </template>
-            <template #footer-extra>
-                <div class="footer-extra">
-                    <p-button style-type="tertiary"
-                              icon-left="ic_copy"
-                              @click="handleClickCopy"
-                    >
-                        {{ $t('ALERT_MANAGER.ALERTS.COPY_ALL') }}
-                    </p-button>
-                </div>
-            </template>
-            <template #confirm-button>
-                {{ $t('ALERT_MANAGER.ALERTS.OK') }}
-            </template>
-        </p-button-modal>
+        <alert-detail-tabs-timeline-modal v-if="state.modalVisible"
+                                          :visible.sync="state.modalVisible"
+                                          :type="state.modalType"
+                                          :history="state.selectedItem"
+        />
     </section>
 </template>
