+++ apps/web/src/services/alert-manager/components/ServiceDetailTabsSettingsEventRuleFormCard.vue
@@ -49,8 +49,6 @@ const allReferenceGetters = allReferenceStore.getters;
 const serviceDetailPageStore = useServiceDetailPageStore();
 const serviceDetailPageState = serviceDetailPageStore.state;
 
-const emit = defineEmits<{(e: 'confirm'): void}>();
-
 const storeState = reactive({
     serviceId: computed<string>(() => serviceDetailPageState.serviceInfo.service_id),
     isEventRuleEditMode: computed<boolean>(() => serviceDetailPageState.isEventRuleEditMode),
@@ -190,9 +188,7 @@ const handleAddButton = async () => {
         await serviceDetailPageStore.fetchEventRuleList({
             service_id: storeState.serviceId,
         });
-        if (storeState.isEventRuleEditMode) {
-            emit('confirm');
-        } else {
+        if (!storeState.isEventRuleEditMode) {
             await replaceUrlQuery({
                 webhookId: response.webhook_id || 'global',
                 eventRuleId: response.event_rule_id,
