+++ apps/web/src/services/alert-manager/constants/alert-table-constant.ts
@@ -45,3 +45,10 @@ export const ALERT_EXCEL_FIELDS: ExcelDataField[] = [
     { key: 'triggered_by', name: 'Triggered by' },
     { key: 'created_at', name: 'Created' },
 ];
+
+export const ALERT_CHANNEL_TABLE_FIELDS: DataTableFieldType[] = [
+    { name: 'type', label: 'Type' },
+    { name: 'name', label: 'Name' },
+    { name: 'state', label: 'State' },
+    { name: 'error_message', label: 'Message' },
+];
