+++ apps/web/src/services/cost-explorer/stores/cost-analysis-page-store.ts
@@ -35,8 +35,6 @@ import type {
     DisplayDataType,
 } from '@/services/cost-explorer/types/cost-explorer-query-type';
 
-
-
 const getRefinedFilters = (consoleFilters?: ConsoleFilter[]): Record<string, string[]> => {
     if (!consoleFilters || isEmpty(consoleFilters)) return {};
     const result: Record<string, string[]> = {};
